This should completely elimiate runtime branching, and remove the buf.Rewind(2) in cases where the last field isn't "omitempty".
The branch can be removed because we now insert an extra space when starting an object. If no fields are written, buf.Rewind(1) will not just delete that space.

The result is code that is much cleaner and faster when there are omitempties in a struct, and as efficient as handwritten code.