Fix usage of vendored type
Package path is needed to be changed before importing it.
The package path of package github.com/a/b if it is in the
vendor directory will be $GOPATH/my/project/vendor/github.com/a/b.
That can't be used in the import statement.
This change trims everything before and including a path containing
a '/vendor/' in it.

Fixes #214